src/main_service/requirements.txt
aas-core3.0
pandas
joblib
fastapi
uvicorn
docker